# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.1.1",
    "title": "Student Rest API with Swagger",
    "description": "Documentation generated by the <b><i>swagger-autogen</i></b> module"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Student",
      "description": "Student data"
    },
    {
      "name": "Teacher",
      "description": "Teacher data"
    },
    {
      "name": "Lecture",
      "description": "Lecture data"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application.json"
  ],
  "produces": [
    "application.json"
  ],
  "paths": {
    "/students/": {
      "get": {
        "tags": [
          "Student"
        ],
        "description": "GET all students with their courses and teachers",
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Student"
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "An array of the all students"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Student"
        ],
        "description": "Register a new student",
        "parameters": [
          {
            "name": "newStudent",
            "in": "body",
            "description": "New student data",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "firstname": {
                  "type": "string",
                  "example": "Tom"
                },
                "lastname": {
                  "type": "string",
                  "example": "Clement"
                },
                "classname": {
                  "type": "string",
                  "example": "B2"
                },
                "age": {
                  "type": "number",
                  "example": 7
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "New registered student",
            "schema": {
              "$ref": "#/definitions/Student"
            }
          },
          "412": {
            "description": "Precondition Failed"
          }
        }
      }
    },
    "/students/name/{name}": {
      "get": {
        "tags": [
          "Student"
        ],
        "description": "GET student by firstname or lastname with their courses and teachers",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Student name existed in the database"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Student"
            },
            "description": "Student requested by id"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/students/id/{id}": {
      "get": {
        "tags": [
          "Student"
        ],
        "description": "GET student by id with their courses and teachers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Student name existed in the database"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Student"
            },
            "description": "Student requested by id"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/students/{id}": {
      "put": {
        "tags": [
          "Student"
        ],
        "description": "Update a stundet data",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Student id existed in the database"
          },
          {
            "name": "newStudent",
            "in": "body",
            "description": "New data for updating an existed student",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "firstname": {
                  "type": "string",
                  "example": "Tom"
                },
                "lastname": {
                  "type": "string",
                  "example": "Clement"
                },
                "classname": {
                  "type": "string",
                  "example": "B2"
                },
                "age": {
                  "type": "number",
                  "example": 7
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Updated student",
            "schema": {
              "$ref": "#/definitions/Student"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Student"
        ],
        "description": "DELETE a student by id ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Student id existed in the database"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/teachers/": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "description": "GET all teachers",
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Teacher"
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "An array of the all teachers"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Teacher"
        ],
        "description": "Register a new teacher",
        "parameters": [
          {
            "name": "newTeacher",
            "in": "body",
            "description": "New teacher data",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "firstname": {
                  "type": "string",
                  "example": "David"
                },
                "lastname": {
                  "type": "string",
                  "example": "Lopez"
                },
                "lecture_id": {
                  "type": "number",
                  "example": 2
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "New registered teacher",
            "schema": {
              "$ref": "#/definitions/Teacher"
            }
          },
          "412": {
            "description": "Precondition Failed"
          }
        }
      }
    },
    "/teachers/name/{name}": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "description": "GET teacher by firstname or lastname",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Teacher name existed in the database"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Teacher"
            },
            "description": "Teacher requested by id"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/teachers/id/{id}": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "description": "GET teacher by id. If the theacher has no lecture associated then there is no result shown since it is inner join on Lecture table. /name/{teacher name} request shows all teachers regardless of lecture_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Teacher name existed in the database"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Teacher"
            },
            "description": "Teacher requested by id"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/teachers/{id}": {
      "put": {
        "tags": [
          "Teacher"
        ],
        "description": "Update a teacher data",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Teacher id existed in the database"
          },
          {
            "name": "newTeacher",
            "in": "body",
            "description": "New data for updating an existed teacher",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "firstname": {
                  "type": "string",
                  "example": "David"
                },
                "lastname": {
                  "type": "string",
                  "example": "Lopez"
                },
                "lecture_id": {
                  "type": "number",
                  "example": 2
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Updated teacher",
            "schema": {
              "$ref": "#/definitions/Teacher"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Teacher"
        ],
        "description": "DELETE a teacher by id ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Teacher id existed in the database"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/lectures/": {
      "get": {
        "tags": [
          "Lecture"
        ],
        "description": "GET all lectures with responsible people",
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Lecture"
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "An array of the all lectures"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Lecture"
        ],
        "description": "Register a new lecture",
        "parameters": [
          {
            "name": "newLecture",
            "in": "body",
            "description": "New lecture data",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Mathematics"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "New registered lecture",
            "schema": {
              "$ref": "#/definitions/Lecture"
            }
          },
          "412": {
            "description": "Precondition Failed"
          }
        }
      }
    },
    "/lectures/name/{name}": {
      "get": {
        "tags": [
          "Lecture"
        ],
        "description": "GET lecture by firstname or lastname",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Lecture name existed in the database"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Lecture"
            },
            "description": "Lecture requested by id"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/lectures/id/{id}": {
      "get": {
        "tags": [
          "Lecture"
        ],
        "description": "GET lecture by id. If the lecture has no teacher associated then there is no result shown since it is inner join on Teacher table. /name/{lecture name} request shows all lectures regardless of teachers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Lecture name existed in the database"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Lecture"
            },
            "description": "Lecture requested by id"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/lectures/{id}": {
      "put": {
        "tags": [
          "Lecture"
        ],
        "description": "Update a stundet data",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Lecture id existed in the database"
          },
          {
            "name": "newLecture",
            "in": "body",
            "description": "New data for updating an existed lecture",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Mathematics"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Updated lecture",
            "schema": {
              "$ref": "#/definitions/Lecture"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Lecture"
        ],
        "description": "DELETE a lecture by id ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Lecture id existed in the database"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "Student": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "firstname": {
          "type": "string",
          "example": "Tom"
        },
        "lastname": {
          "type": "string",
          "example": "Clement"
        },
        "classname": {
          "type": "string",
          "example": "A1"
        },
        "age": {
          "type": "number",
          "example": 7
        },
        "Courses": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "History"
              }
            }
          }
        },
        "Teachers": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "firstname": {
                "type": "string",
                "example": "Noah"
              },
              "lastname": {
                "type": "string",
                "example": "François"
              }
            }
          }
        }
      }
    },
    "Teacher": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "firstname": {
          "type": "string",
          "example": "David"
        },
        "lastname": {
          "type": "string",
          "example": "Möller"
        },
        "Specialty": {
          "type": "string",
          "example": "Mathematics"
        }
      }
    },
    "Lecture": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Mathematics"
        },
        "The Responsible": {
          "type": "object",
          "properties": {
            "firstname": {
              "type": "string",
              "example": "Emma"
            },
            "lastname": {
              "type": "string",
              "example": "Durand"
            }
          }
        }
      }
    }
  }
}